Bug 36055: Simplify MARC21 Fast Add framework
This patch removes or hides a lot of unecessary fields from the
MARC21 Fast Add framework, notably subfields $6 and $8 which are
used only in advanced cataloging. Fast Add is meant to add an item
fast, while the patron is waiting, in order to check it out to them.
There is no need to enter details, only what is necessary to
identify and check out the item.

To test:
1. Go to Circulation > Fast cataloging
   --> Notice all the fields present
2. Fill out mandatory fields (000, 008, 245$a, 942$c)
3. Click 'Save'
   --> Notice all the item fields present
4. Apply patch and reset
5. Redo steps 1-3
   --> There should be a lot less fields

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Esther <esther@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>